add (passing) tests for field ordering
